Refactor if-else switch to python dictionary switch
Created "winStates" dictionary, this way if there is not a tie we look
directly for one of the moves and see whether the computer move is a
part of the list. If it is, the player wins. If it is not, the
computer beat the player.

This method is less code (12 lines added vs. 36 deleted) and should be
a little more robust with the user interface (i.e. the player cannot
give an invalid move like "ock" or "0" when asked for input).